+++ packages/myst-common/package.json
@@ -30,7 +30,9 @@
     "vfile-message": "^3.0.0"
   },
   "devDependencies": {
-    "myst-spec-ext": "^1.1.22",
-    "unist-builder": "3.0.0"
+    "myst-spec-ext": "^1.1.21",
+    "unist-builder": "3.0.0",
+    "@jupyterlab/nbformat": "^3.5.2",
+    "@lumino/coreutils": "^2.0.0"
   }
 }
