+++ packages/myst-execute/src/index.ts
@@ -3,4 +3,3 @@ export type { JupyterServerSettings } from './manager.js';
 export { findExistingJupyterServer, launchJupyterServer } from './manager.js';
 export { LocalDiskCache } from './cache.js';
 export type { ICache } from './cache.js';
-export type { IExpressionResult, IExpressionError, IExpressionOutput } from './types.js';
